Consolidate benchmark code to eliminate duplication
Extract shared benchmark logic into benches/shared/{bench_full,bench_fast}.rs
templates, reducing each per-ciphersuite benchmark file to a 2-line stub
that imports suite types and includes the template. Normalizes Scalar::from
to use u128 uniformly across all ciphersuites.
